pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docudialogue" # This is how you'll import it: import docudialogue
version = "0.1.0"
requires-python = ">=3.8"

dependencies = [
    "python-dotenv",
    "haystack-ai",
    "pydantic",
    "openai",
    "markdown-it-py",
    "mdit_plain",
    "pypdf",
    "neo4j",
    "igraph>=0.11.8",
    "scipy>=1.10.1",
    "matplotlib>=3.7.5",
    "leidenalg>=0.10.2",
    "ipykernel>=6.29.5",
]

[tool.setuptools.packages.find]
where = ["docudialogue/src"]
include = ["docudialogue*"] 
